[Chore #160311539] refactor MyArticlesPage
